Add Personage List on 3.6.10.10 and beginning migrating
